fixes due to lints being removed
slightly bigger impact than expected, only generate req/responses
for broker messages (brokerheartbeat, which is a controller message
has an unused tag_message, which is really a bug... but we don't
use controller messages, so disabled until (if?) we need them).

broker registration, removed listeners and features as they
originate from controller messages. As we have stateless brokers,
broker registrion is a bit mute (all brokers register the same
information). pg storage engine does need to "register" the cluster,
which broker registration is being used for.. which ultimately
needs to go, but not as part of this branch